Moves custom live data classes into own package
